scheduler/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
claid==0.6.4
click==8.1.7
Flask==3.1.0
Flask-APScheduler==1.13.1
grpcio==1.59.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.25.0
packaging==24.2
pandas==2.2.3
protobuf==4.25.3
psutil==6.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.1.3
claid==0.6.4
pandas==2.2.3
numpy==1.25.0

